modules/dev/adhd_mcp/pyproject.toml
[project]
name = "adhd-mcp"
version = "0.0.1"
description = "ADHD Framework MCP: adhd_mcp"
requires-python = ">=3.11"
dependencies = [
    "logger-util",
    "config-manager",
    "modules-controller-core",
    "module-creator-core",
    "github-api-core",
    "creator-common-core",
    "exceptions-core",
    "instruction-core",
    "pyyaml>=6.0",
    "mcp>=1.1.0",
]

[project.urls]
Repository = "https://github.com/AI-Driven-Highspeed-Development/adhd_mcp.git"

[tool.adhd]
layer = "dev"
mcp = true

[tool.uv.sources]
logger-util = { workspace = true }
config-manager = { workspace = true }
modules-controller-core = { workspace = true }
module-creator-core = { workspace = true }
github-api-core = { workspace = true }
creator-common-core = { workspace = true }
exceptions-core = { workspace = true }
instruction-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build]
dev-mode-dirs = [".."]
